read the words

+++ djangotcha/management/commands/assign_targets.py
@@ -1,7 +1,8 @@
-from django.core.management.base import BaseCommand, CommandError
+import os
+import random
+from django.core.management.base import BaseCommand
 from djangotcha.models import Person, Assassination
 from django.conf import settings
-import random
 
 
 class Command(BaseCommand):
@@ -9,11 +10,15 @@ class Command(BaseCommand):
     help = 'Assign random targets to everyone'
 
     def handle(self, *args, **options):
+        if True:
+            read()
+            return
+
         Assassination.objects.all().delete()
 
         persons = Person.objects.filter(user__is_superuser=False).order_by('?')
 
-        secret_words = settings.SECRET_WORDS
+        secret_words = read()
         random.shuffle(secret_words)
 
         i = 0
@@ -23,15 +28,23 @@ class Command(BaseCommand):
             else:
                 target = persons[i-1]
 
-            secret_word = secret_words[ i % len(secret_words) ]
-
             self.stdout.write('assigned: %s #\'== ~> %s' % (p,target))
 
             p.target = target
-            p.secret_word = secret_word
+            p.secret_word = secret_words[i]
             p.is_killed = False
             p.date_killed = None
             p.save()
             i += 1
 
         self.stdout.write('Successfully assigned targets')
+
+
+def read():
+    current = os.path.dirname(os.path.abspath(__file__))
+    path = os.path.join(current, 'words')
+    fh = open(path, 'r')
+
+    words = [line.strip() for line in fh]
+    return words
+
